requirements.txt
torch>=1.0.0
requests>=2.25.0
Pillow>=8.0.0
numpy>=1.19.0

